requirements.txt
langchain==0.2.16
langchain-community==0.2.11
langchain-openai==0.1.21
langchain-text-splitters==0.2.2
faiss-cpu==1.12.0
tiktoken==0.11.0
pypdf==4.3.1
python-dotenv==1.0.1
streamlit==1.37.1
